Improve CommandMessageReceivedEvent#isBotMentioned()
Previously, the method would always return true due to an incorrect `>= 1` statement.
It also would ignore the @ in the command name e.g. in `/command@botname`, it would
ignore the fact that the bot `@botname` was the one targetted, and would always return
true. This suggested to the API user that their bot was intended to use the command.

Now, this conditioning is fixed to ensure the method only returns true when the bot is
directly targeted.

In addition, when the bot is messaged in a private chat (not a group, supergroup,
or channel), the method will return true. This is because the bot is being implicitly
mentioned, as in the chat there exists only the user and the bot.